Move tests to the prompt executor module
